+++ NEWS.md
@@ -6,6 +6,8 @@ New features:
  * New authentication mechanism: checks in advance if credentials are provided
    in order to do a single request to redis to retrieve the required database
    connection parameters.
+ * Retrieves OAuth hash values using new cartodb-redis method so it will reuse
+   the redis pool from cartodb-redis instead of using just one pool for oauth.
 
 
 1.12.1 - 2014-08-05